requirement.txt
google-generativeai==0.5.2
playsound==1.2.2
gTTS==2.5.1
SpeechRecognition==3.10.3
python-dotenv==1.0.1